These is the first attempt to optimize this kernel. Couple of things we've tried:
1. numNeighbors is an array of size num_particles, and keeps the number
of neighbors. It is stored in global memory. Somehow that is repeated
accesses to increment the counter overtime. We move the aggregation into
a local register before finally commiting to a singel global memory
write

2. 2 kernels are launched for fuild simulation solver: findLambdasD and
solveFluidsD. They are in a producer/consumer relationship, former
procuces lambdas and neighborhood information and latter consumes it. We
perform kernel fusion of the 2, reducing a lot of common code such as
loading particle positions, checking, etc, and use a syncthreads instead
of global barrier. We do not that this might cause numerical differences
to the simulation but an visual inspection did not seem to find anything
wierd.