This commit is another attempt to optimize findLambdasD+solveFluidsD
Things that worked:
1. Making "rad" a constant instead of calculated (since it always 4 as
far as this simulation is concerned)

2. Completely eliminate numNeighbors as its solely used in this kernel
and now that kernel fusion is performed, no global variable is needed to
pass information around.

(from previous commit): we cached read to oldPos of the current particle
that used to be present to both kernels.

Things that did not work
Loop unroll: somehow unrolling the "rad" loop after we made it a
constant hurt performance by 2-3X

Caching read of neighbor positions: We hoped to cache the position of
neighbors, instead of cacheing thier index into the global array oldPos.
This seemed to cause more harm than good, making it 2X slower :(

Finally we are making the "scale" of the system into 6, which now stands
around 30 FPS on my local desktop. The system should run up to 60FPS so
there is still rooms for improvement, but this makes the simulation look
somewhat smooth.